Restructured pk-pk reporting
Modified how pk-pk reporting works - there's a designated register
for it now. The time-series trace is still stored in memory and can
be accessed however the user sees fit (it occupies the same place in
the address space as it did previously).

The frequency is no longer encoded in the pk-pk amplitude because
it is specified by the user (so they definitely know what it is).

The code to write the virus bitmask was cleaned up so that it
takes up less space and is somewhat more adjustable.

The .gitignore file was updated using an online template.